LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/parsing/column-rule-width-computed-expected.txt


PASS Property column-rule-width value 'calc(10px + 0.5em)' computes to '30px' 
PASS Property column-rule-width value 'calc(10px - 0.5em)' computes to '0px' 
PASS column-rule-width is 0 when column-rule-style is none or hidden 
